requirements/production.txt
Django>=1.9.5, <1.10
psycopg2==2.6.1
prettyconf==1.2.3
dj-database-url==0.4.1
djangorestframework==3.3.3
dj-email-url==0.0.4
django-rest-swagger==0.3.5
django-mptt==0.8.3
drf-extensions==0.2.8
